refactor: move quiz into a stand alone component
